+++ lib/smart_form_fields.dart
@@ -5,6 +5,8 @@ export 'src/form/smart_form.dart' show SmartForm, SmartFormState;
 export 'src/form/smart_form_controller.dart' show SmartFormController;
 export 'src/form/smart_form_key.dart' show SmartFormKey;
 export 'src/form/smart_form_result.dart' show SmartFormResult;
+export 'src/theme/smart_form_theme.dart'
+    show SmartFormTheme, SmartFormThemeData;
 export 'src/fields/smart_field_controller.dart' show SmartFieldController;
 export 'src/fields/smart_date_field.dart'
     show SmartDateField, SmartDateFormatter;